Move localization.js to app
Since it is no longer used in core. Also splits localization tests into
a separate file.
